+++ python_scripts/task6/apicall/apicall.py
@@ -1,84 +1,7 @@
 #!/usr/bin/env python3
 
-"""
-Practical tasks (Part 2)
-
-Create a script that uses the Survey Monkey (https://www.surveymonkey.com) service,
-creates a survey, and sends an invitation to go through it.
-
-
-PREREQUISITES:
-
-Sign up at https://www.surveymonkey.com 
-
-Create a draft application at https://developer.surveymonkey.com   
-
-No need to deploy your application.
-It's just for testing. Do not forget to set permissions for your application.
-
-After creating a draft application you will obtain an ACCESS_TOKEN,
-which is needed to do API requests from your script.
-
-
-REQUIREMENTS:
-
-The script should accept a JSON file with questions for the survey,
-and a text file with a list of email addresses.
-
-The structure of a JSON file with questions:
-
-{
-
-   "Survey_Name": {
-
-      "Page_Name": {
-
-          "Question1_Name": {
-
-              "Description" : "Description of question",
-
-              "Answers" : [
-
-                  "Answer1",
-
-                  "Answer2",
-
-                  "Answer3"
-
-              ]
-
-          },
-
-          "Question2_Name": {
-
-              "Description" : "Description of question",
-
-              "Answers" : [
-
-                  "Answer1",
-
-                  "Answer2",
-
-                  "Answer3"
-
-              ]
-
-          }
-
-          . . .
-
-      }
-
-   }
-
-}
-
-There should be at least 3 questions and 2 recipients.
-"""
-
 import requests
 import json
-import emailbot
 
 ACCESS_TOKEN="cMFftolwVdUKH2zJtk2zDOMeAKkYgQk3RxT7usJRmmEjnrLw8YQ52QIhm7OhjibN.Mz86CNoWe1tq06i52S982BC5bqZrzuIykpxGsAxAa2vsOUmnD3vqJUJwObv4ten"
 
@@ -165,10 +88,3 @@ def parse_reps():
         reps.append(line.replace("\n",""))
   return reps
 
-
-if __name__ == '__main__':
-  webllink=create_weblink(postSurvey())
-  email_list=parse_reps()
-  for email in email_list:
-     #sends emails with survey
-     emailbot.gmail_send_message(content=webllink,email=email)
